+++ WordPress/src/main/java/org/wordpress/android/ui/domains/management/DomainManagementActivity.kt
@@ -12,6 +12,7 @@ import kotlinx.coroutines.flow.onEach
 import org.wordpress.android.ui.ActivityLauncher
 import org.wordpress.android.ui.domains.management.DomainManagementViewModel.ActionEvent
 import org.wordpress.android.ui.domains.management.details.DomainManagementDetailsActivity
+import org.wordpress.android.ui.themes.M3Theme
 import org.wordpress.android.util.extensions.setContent
 
 @AndroidEntryPoint
