+++ WordPress/src/main/java/org/wordpress/android/ui/domains/management/DomainStatusRow.kt
@@ -31,6 +31,7 @@ import org.wordpress.android.R
 import org.wordpress.android.fluxc.network.rest.wpcom.site.DomainStatus
 import org.wordpress.android.fluxc.network.rest.wpcom.site.StatusType
 import org.wordpress.android.ui.domains.management.composable.PendingGhostStrip
+import org.wordpress.android.ui.themes.M3Theme
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
 import java.time.format.FormatStyle
