+++ WordPress/src/main/java/org/wordpress/android/ui/domains/management/composable/PendingGhostStrip.kt
@@ -17,7 +17,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.unit.Dp
-import org.wordpress.android.ui.domains.management.ghost
+import org.wordpress.android.ui.themes.ghost
 
 
 @Composable
