+++ WordPress/src/main/java/org/wordpress/android/ui/domains/management/composable/PrimaryButton.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import org.wordpress.android.ui.compose.theme.AppColor
-import org.wordpress.android.ui.domains.management.M3Theme
+import org.wordpress.android.ui.themes.M3Theme
 
 @Composable
 fun PrimaryButton(
