+++ WordPress/src/main/java/org/wordpress/android/ui/domains/management/newdomainsearch/composable/JetpackOutlinedButton.kt
@@ -11,7 +11,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import org.wordpress.android.ui.domains.management.success
+import org.wordpress.android.ui.themes.success
 
 @Composable
 fun JetpackOutlinedButton(
